refactor(switch): consolidate get peer utils
