+++ src/switch/index.js
@@ -8,7 +8,7 @@ const series = require('async/series')
 const Circuit = require('libp2p-circuit')
 const TransportManager = require('./transport')
 const ConnectionManager = require('./connection/manager')
-const getPeerInfo = require('./get-peer-info')
+const { getPeerInfo } = require('../get-peer-info')
 const getDialer = require('./dialer')
 const connectionHandler = require('./connection/handler')
 const ProtocolMuxer = require('./protocol-muxer')
